// src/app/models/unityOfMeasure.d.ts
/**
 * Pick Voice
 * Api para el sistema Pick Voice, plataforma que optimiza los procesos logisticos de un centro de distribucion, implementando tecnologia por voz.
 *
 * OpenAPI spec version: 1.0.0
 * Contact: gabriel.martinez@tau-tech.co
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
export interface UnityOfMeasure {
    code?: string;
    name?: string;
}